# dashboard/tmp/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Scheduler Public API (v1)
  version: 1.3.0
servers:
  - url: 'https://scheduler.example.com'
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/workflow-packages:
    get:
      tags:
        - WorkflowPackages
      summary: List published workflow packages
      operationId: listWorkflowPackages
      parameters:
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
        - name: owner
          in: query
          description: Filter by owner id; use `me` for the caller's id.
          schema:
            type: string
        - name: visibility
          in: query
          description: 'Filter by visibility (private, internal, public).'
          schema:
            type: string
        - name: search
          in: query
          description: 'Full-text search across slug, display name, and summary.'
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPackageList'
  '/api/v1/workflow-packages/{packageId}':
    get:
      tags:
        - WorkflowPackages
      summary: Get a workflow package detail
      operationId: getWorkflowPackage
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPackageDetail'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - WorkflowPackages
      summary: Delete a workflow package
      operationId: deleteWorkflowPackage
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          $ref: '#/components/responses/NotFound'
  '/api/v1/workflow-packages/{packageId}/versions':
    get:
      tags:
        - WorkflowPackages
      summary: List versions for a workflow package
      operationId: listWorkflowPackageVersions
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPackageVersionList'
        '404':
          $ref: '#/components/responses/NotFound'
  '/api/v1/workflow-packages/{packageId}/clone':
    post:
      tags:
        - WorkflowPackages
      summary: Clone a workflow package version into the caller's workspace
      operationId: cloneWorkflowPackage
      parameters:
        - name: packageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowPackageCloneRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowRef'
        '404':
          $ref: '#/components/responses/NotFound'
  '/api/v1/workflows/{workflowId}/publish':
    post:
      tags:
        - WorkflowPackages
      summary: Publish a workflow draft to the Store
      operationId: publishWorkflow
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowPublishRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowPublishResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /api/v1/workers:
    get:
      tags:
        - Workers
      summary: List workers (scheduler view)
      operationId: listWorkers
      parameters:
        - name: queue
          in: query
          required: false
          schema:
            type: string
        - $ref: '#/components/parameters/PageLimit'
        - $ref: '#/components/parameters/PageCursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Worker'
                  nextCursor:
                    type: string
                    nullable: true
  '/api/v1/workers/{workerId}':
    get:
      tags:
        - Workers
      summary: Get worker snapshot
      operationId: getWorker
      parameters:
        - $ref: '#/components/parameters/WorkerId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Worker'
        '404':
          $ref: '#/components/responses/NotFound'
  '/api/v1/workers/{workerId}/commands':
    post:
      tags:
        - Workers
      summary: Enqueue admin command (drain/rebind/pkg.install/pkg.uninstall)
      operationId: sendWorkerCommand
      parameters:
        - $ref: '#/components/parameters/WorkerId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkerCommand'
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommandRef'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkerId:
      name: workerId
      in: path
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        maxLength: 64
      description: 'Optional idempotency key for safe retries; if reused with a different body, return 409'
    PageLimit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    PageCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 'Conflict (e.g., idempotency-key reuse with different body)'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Authenticated but lacks required permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    WorkflowRef:
      type: object
      required:
        - workflowId
      properties:
        workflowId:
          type: string
          format: uuid
...
    Worker:
      type: object
      required:
        - id
        - lastHeartbeatAt
        - queues
      properties:
        id:
          type: string
        hostname:
          type: string
        lastHeartbeatAt:
          type: string
          format: date-time
        queues:
          type: array
          items:
            type: string
        packages:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              version:
                type: string
              status:
                type: string
                enum:
                  - installed
                  - installing
                  - uninstalling
                  - removed
                  - failed
                  - unknown
        meta:
          type: object
          additionalProperties: true
    WorkerCommand:
      type: object
      oneOf:
        - properties:
            type:
              const: drain
          required:
            - type
        - properties:
            type:
              const: rebind
            queue:
              type: string
          required:
            - type
            - queue
        - properties:
            type:
              const: pkg.install
            name:
              type: string
            version:
              type: string
          required:
            - type
            - name
        - properties:
            type:
              const: pkg.uninstall
            name:
              type: string
          required:
            - type
            - name
      additionalProperties: false
    CommandRef:
      type: object
      required:
        - commandId
        - workerId
        - acceptedAt
      properties:
        commandId:
          type: string
        workerId:
          type: string
        acceptedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        requestId:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
    WorkflowPackageVersion:
      type: object
      required:
        - id
        - version
        - publishedAt
      properties:
        id:
          type: string
        version:
          type: string
        changelog:
          type: string
          nullable: true
        publishedAt:
          type: string
          format: date-time
        publisherId:
          type: string
          nullable: true
        previewImage:
          type: string
          nullable: true
          description: Base64-encoded preview of the workflow canvas for this version.
    WorkflowPackageVersionList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowPackageVersion'
    WorkflowPackageSummary:
      type: object
      required:
        - id
        - slug
        - displayName
        - visibility
      properties:
        id:
          type: string
        slug:
          type: string
        displayName:
          type: string
        summary:
          type: string
          nullable: true
        visibility:
          type: string
          description: Visibility policy (private/internal/public).
        tags:
          type: array
          items:
            type: string
          nullable: true
        ownerId:
          type: string
          nullable: true
        ownerName:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        latestVersion:
          $ref: '#/components/schemas/WorkflowPackageVersion'
        previewImage:
          type: string
          nullable: true
          description: Base64-encoded preview from the latest version when available.
    WorkflowPackageDetail:
      allOf:
        - $ref: '#/components/schemas/WorkflowPackageSummary'
        - type: object
          required:
            - versions
          properties:
            versions:
              type: array
              items:
                $ref: '#/components/schemas/WorkflowPackageVersion'
    WorkflowPackageList:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowPackageSummary'
        nextCursor:
          type: string
          nullable: true
    WorkflowPackageCloneRequest:
      type: object
      properties:
        versionId:
          type: string
          description: Explicit version id to clone; defaults to the latest.
        version:
          type: string
          description: Version string to clone when versionId is omitted.
        workflowName:
          type: string
          description: Optional override for the cloned workflow name.
    WorkflowPublishRequest:
      type: object
      required:
        - version
      properties:
        packageId:
          type: string
          description: Existing package id to append a version to.
        slug:
          type: string
          description: Slug used when creating a new package.
        displayName:
          type: string
        summary:
          type: string
          nullable: true
        visibility:
          type: string
          description: Desired visibility (private/internal/public).
        tags:
          type: array
          items:
            type: string
          nullable: true
        version:
          type: string
          description: Semantic version for the published snapshot.
        changelog:
          type: string
          nullable: true
        previewImage:
          type: string
          nullable: true
          description: Base64-encoded preview captured from the workflow canvas.
    WorkflowPublishResponse:
      type: object
      required:
        - packageId
        - versionId
        - version
      properties:
        packageId:
          type: string
        versionId:
          type: string
        version:
          type: string